refactor: rename http to http_repository
